# docs/api/consolidated_api_docs.html/auth-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Auth Service",
    "description": "Authentication and authorization service for Cryptobot",
    "version": "1.0.0"
  },
  "paths": {
...
    "/admin/admin/sessions": {
      "get": {
        "tags": [
          "admin",
          "admin"
        ],
        "summary": "List Active Sessions",
        "description": "List all active sessions (optionally filtered by user)",
        "operationId": "list_active_sessions_admin_admin_sessions_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionOut"
                  },
                  "title": "Response List Active Sessions Admin Admin Sessions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/admin/sessions/terminate": {
      "post": {
        "tags": [
          "admin",
          "admin"
        ],
        "summary": "Terminate Session",
        "description": "Terminate a session by ID",
        "operationId": "terminate_session_admin_admin_sessions_terminate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionTerminate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/admin/admin/sessions/suspicious": {
      "post": {
        "tags": [
          "admin",
          "admin"
        ],
        "summary": "Report Suspicious Activity",
        "description": "Report suspicious session activity",
        "operationId": "report_suspicious_activity_admin_admin_sessions_suspicious_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspiciousActivityCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "session_token": {
            "type": "string",
            "title": "Session Token"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "is_suspicious": {
            "type": "boolean",
            "title": "Is Suspicious"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_activity": {
            "type": "string",
            "format": "date-time",
            "title": "Last Activity"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_token",
          "user_id",
          "is_active",
          "is_suspicious",
          "created_at",
          "last_activity"
        ],
        "title": "SessionOut"
      },
      "SessionTerminate": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Session terminated"
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "SessionTerminate"
      },
      "SuspiciousActivityCreate": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "activity_type": {
            "type": "string",
            "title": "Activity Type"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "activity_type",
          "severity"
        ],
        "title": "SuspiciousActivityCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}